# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LightTickets API",
    "version": "1.0.0",
    "description": "LightTickets API 文档"
  },
  "servers": [
    {
      "url": "http://localhost:23320",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Server-Key"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lt_refresh_token"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/setup/site-config": {
      "get": {
        "summary": "获取站点配置",
        "tags": [
          "Setup"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "isSetup": {
                          "type": "boolean"
                        },
                        "requireLogin": {
                          "type": "boolean"
                        },
                        "allowWebRegister": {
                          "type": "boolean"
                        },
                        "allowMcRegister": {
                          "type": "boolean"
                        },
                        "passwordResetEnabled": {
                          "type": "boolean",
                          "description": "True only when SMTP is usable and siteUrl is a canonical HTTPS origin"
                        },
                        "registrationEmailVerificationEnabled": {
                          "type": "boolean",
                          "description": "True when SMTP is usable"
                        },
                        "siteName": {
                          "type": "string"
                        },
                        "siteUrl": {
                          "type": "string",
                          "nullable": true,
                          "format": "uri",
                          "description": "Canonical public HTTP(S) origin without credentials, path, query, or fragment"
                        },
                        "footerContent": {
                          "type": "string",
                          "nullable": true
                        },
                        "defaultLanguage": {
                          "type": "string"
                        },
                        "turnstile": {
                          "type": "object",
                          "properties": {
                            "enabled": {
                              "type": "boolean"
                            },
                            "siteKey": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "enabled",
                            "siteKey"
                          ]
                        },
                        "federatedAuthProviders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "slug": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "iconUrl": {
                                "type": "string",
                                "nullable": true
                              },
                              "allowRegistration": {
                                "type": "boolean"
                              }
                            },
                            "required": [
                              "slug",
                              "name",
                              "iconUrl",
                              "allowRegistration"
                            ]
                          }
                        }
                      },
                      "required": [
                        "isSetup",
                        "requireLogin",
                        "allowWebRegister",
                        "allowMcRegister",
                        "passwordResetEnabled",
                        "registrationEmailVerificationEnabled",
                        "siteName",
                        "siteUrl",
                        "footerContent",
                        "defaultLanguage",
                        "turnstile",
                        "federatedAuthProviders"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/setup": {
      "post": {
        "summary": "执行初始化设置",
        "tags": [
          "Setup"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "db": {
                    "type": "object",
                    "properties": {
                      "provider": {
                        "type": "string",
                        "enum": [
                          "sqlite",
                          "mysql"
                        ]
                      },
                      "host": {
                        "type": "string"
                      },
                      "port": {
                        "type": "integer",
                        "minimum": 0,
                        "exclusiveMinimum": true
                      },
                      "username": {
                        "type": "string"
                      },
                      "password": {
                        "type": "string"
                      },
                      "database": {
                        "type": "string"
                      },
                      "args": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "provider"
                    ],
                    "additionalProperties": false
                  },
                  "admin": {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string",
                        "format": "email"
                      },
                      "password": {
                        "type": "string",
                        "minLength": 6
                      },
                      "username": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 30
                      }
                    },
                    "required": [
                      "email",
                      "password",
                      "username"
                    ]
                  },
                  "site": {
                    "type": "object",
                    "properties": {
                      "siteName": {
                        "type": "string"
                      },
                      "siteUrl": {
                        "type": "string",
                        "description": "Canonical public HTTP(S) origin; an empty string clears the optional setting"
                      },
                      "defaultLanguage": {
                        "type": "string"
                      }
                    }
                  },
                  "mc": {
                    "type": "object",
                    "properties": {
                      "defaultServerName": {
                        "type": "string"
                      }
                    }
                  },
                  "storage": {
                    "type": "object",
                    "properties": {
                      "driver": {
                        "type": "string",
                        "enum": [
                          "local",
                          "s3"
                        ]
                      },
                      "s3": {
                        "type": "object",
                        "properties": {
                          "endpoint": {
                            "type": "string"
                          },
                          "region": {
                            "type": "string"
                          },
                          "bucket": {
                            "type": "string"
                          },
                          "accessKeyId": {
                            "type": "string"
                          },
                          "secretAccessKey": {
                            "type": "string"
                          },
                          "forcePathStyle": {
                            "type": "boolean"
                          },
                          "presignExpiry": {
                            "type": "integer",
                            "minimum": 0,
                            "exclusiveMinimum": true
                          }
                        }
                      }
                    },
                    "required": [
                      "driver"
                    ]
                  }
                },
                "required": [
                  "db",
                  "admin"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {},
                          "additionalProperties": {
                            "nullable": true
                          }
                        },
                        {
                          "type": "array",
                          "items": {
                            "nullable": true
                          }
                        },
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "nullable": true
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/setup/settings/mail/test": {
      "post": {
        "summary": "测试 SMTP 连接",
        "tags": [
          "Setup"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mail": {
                    "type": "object",
                    "properties": {
                      "enabled": {
                        "type": "boolean"
                      },
                      "host": {
                        "type": "string"
                      },
                      "port": {
                        "type": "integer",
                        "minimum": 0,
                        "exclusiveMinimum": true
                      },
                      "secure": {
                        "type": "boolean"
                      },
                      "username": {
                        "type": "string",
                        "nullable": true
                      },
                      "password": {
                        "type": "string",
                        "nullable": true
                      },
                      "fromName": {
                        "type": "string"
                      },
                      "fromAddress": {
                        "anyOf": [
                          {
                            "type": "string",
                            "format": "email"
                          },
                          {
                            "type": "string",
                            "enum": [
                              ""
                            ]
                          }
                        ]
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "success",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}